Make queue.coffee PriorityQueue-based.
[refs #17]

That means we're ditching Redis for this. In the future, url-fetcher can
route messages itself if it needs to share the load.

Kue was too much overhead and complexity in this case: it's designed to
handle concurrent clients but it's bad at scheduling jobs for the
future. We need scheduling more than we need concurrency.