Make restart seamless and throttle UrlTaskQueue
This solves two problems:

1. On restart, the url-fetcher would initiate a new fetch of everything.
   That's a lot of things.
2. We never had a guarantee that we'd perform a maximum number of, say,
   Facebook popularity requests per minute. Now we do.

This brings up the possibility that we'll get so many URLs the queues
will completely fill -- as in, we may be holding more than two hours'
worth of Facebook requests. That happens at about 11,000 URLs. At that
point, this solution will still be as fair as possible: it will run the
earliest-scheduled fetches first. (When we get to this problem, the
first step is to track the difference between the scheduled time and the
actual running time of each task.)

[closes #38]