Choose number of invocations based on time gap.
Before, we were deciding when the next invocation should be based on how
many invocations there have been in the past. That's correct, but it
won't work with our old data because we used to fetch every hour. Now,
we decide the next invocation based on how many invocations there
_should_ have been during the time the URL has been in our system.

Three months from now, we should probably revert and actually use the
count. It's simpler.