Make fetch_handler ignore manual UrlVersions
This is great because it makes it makes fetches deterministic: given the
same sequence of UrlGets arriving in order, the outcome will always be
the same.

It's also easy for end-users to understand. If you enter a manual
UrlVersion, it may be because the automatic one was wrong. You don't
want a new automatic one inserted. [refs #37]

[refs #90]